Updated front-page.php with part 1 of X of Adding ShareButton
"GitHub Collab 002" is "adding a SHARE BUTTON" (share to social media, button). (More info found on the wiki there)

This 'commit' is to add the umiumiumi controlled-bug/controlled-error in two places.

Please be sure to remove the umiumi error that appears to the left and to the right of 'ID' once this test is done.